common: Separate locking functions from fi.h into own header file
Move the locking calls from fi.h to fi_lock.h.  This is to cleanup
the fi.h header and make it easier to find the necessary definitions.
